refactor: remove deprecated v1 management endpoints
