+++ extensions/common/api/management-api-configuration/src/main/java/org/eclipse/edc/connector/api/management/configuration/ManagementApiConfigurationExtension.java
@@ -28,7 +28,6 @@ import org.eclipse.edc.web.spi.configuration.WebServiceConfigurer;
 import org.eclipse.edc.web.spi.configuration.WebServiceSettings;
 import org.eclipse.edc.web.spi.provider.ObjectMapperProvider;
 
-import static java.lang.String.format;
 import static org.eclipse.edc.spi.CoreConstants.JSON_LD;
 
 /**
@@ -41,35 +40,17 @@ public class ManagementApiConfigurationExtension implements ServiceExtension {
 
     public static final String NAME = "Management API configuration";
 
-    private static final String DEPRECATED_MANAGEMENT_SETTINGS_GROUP_SUFFIX = "data";
-    private static final String DATA_MANAGEMENT_CONTEXT_ALIAS = "management";
-    private static final int DEFAULT_DATA_MANAGEMENT_API_PORT = 8181;
-    private static final String DEPRECATED_DEFAULT_DATA_MANAGEMENT_API_CONTEXT_PATH = "/api";
-    private static final String DEFAULT_DATA_MANAGEMENT_API_CONTEXT_PATH = "/api/v1/management";
+    private static final String MANAGEMENT_CONTEXT_ALIAS = "management";
+    private static final int DEFAULT_MANAGEMENT_API_PORT = 8181;
+    private static final String DEFAULT_MANAGEMENT_API_CONTEXT_PATH = "/api/v1/management";
 
     public static final String WEB_SERVICE_NAME = "Management API";
 
-    /**
-     * This deprecation is used to permit a softer transition from the deprecated `web.http.data` config group to the
-     * current `web.http.management`
-     *
-     * @deprecated "web.http.management" config should be used instead of "web.http.data"
-     */
-    @Deprecated(since = "milestone8")
-    public static final WebServiceSettings DEPRECATED_SETTINGS = WebServiceSettings.Builder.newInstance()
-            .apiConfigKey("web.http." + DEPRECATED_MANAGEMENT_SETTINGS_GROUP_SUFFIX)
-            .contextAlias(DATA_MANAGEMENT_CONTEXT_ALIAS)
-            .defaultPath(DEPRECATED_DEFAULT_DATA_MANAGEMENT_API_CONTEXT_PATH)
-            .defaultPort(DEFAULT_DATA_MANAGEMENT_API_PORT)
-            .useDefaultContext(true)
-            .name(WEB_SERVICE_NAME)
-            .build();
-
     public static final WebServiceSettings SETTINGS = WebServiceSettings.Builder.newInstance()
-            .apiConfigKey("web.http." + DATA_MANAGEMENT_CONTEXT_ALIAS)
-            .contextAlias(DATA_MANAGEMENT_CONTEXT_ALIAS)
-            .defaultPath(DEFAULT_DATA_MANAGEMENT_API_CONTEXT_PATH)
-            .defaultPort(DEFAULT_DATA_MANAGEMENT_API_PORT)
+            .apiConfigKey("web.http." + MANAGEMENT_CONTEXT_ALIAS)
+            .contextAlias(MANAGEMENT_CONTEXT_ALIAS)
+            .defaultPath(DEFAULT_MANAGEMENT_API_CONTEXT_PATH)
+            .defaultPort(DEFAULT_MANAGEMENT_API_PORT)
             .useDefaultContext(true)
             .name(WEB_SERVICE_NAME)
             .build();
@@ -96,18 +77,7 @@ public class ManagementApiConfigurationExtension implements ServiceExtension {
 
     @Override
     public void initialize(ServiceExtensionContext context) {
-        WebServiceSettings settings;
-        var config = context.getConfig();
-        if (config.hasPath(DEPRECATED_SETTINGS.apiConfigKey()) && !config.hasPath(SETTINGS.apiConfigKey())) {
-            settings = DEPRECATED_SETTINGS;
-            context.getMonitor().warning(
-                    format("Deprecated settings group %s is being used for Management API configuration, please switch to the new group %s",
-                            DEPRECATED_SETTINGS.apiConfigKey(), SETTINGS.apiConfigKey()));
-        } else {
-            settings = SETTINGS;
-        }
-
-        var webServiceConfiguration = configurator.configure(context, webServer, settings);
+        var webServiceConfiguration = configurator.configure(context, webServer, SETTINGS);
 
         context.registerService(ManagementApiConfiguration.class, new ManagementApiConfiguration(webServiceConfiguration));
         webService.registerResource(webServiceConfiguration.getContextAlias(), new AuthenticationRequestFilter(authenticationService));
