+++ extensions/control-plane/api/management-api/contract-agreement-api/src/main/java/org/eclipse/edc/connector/api/management/contractagreement/ContractAgreementApi.java
@@ -19,9 +19,10 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
+import jakarta.json.JsonObject;
 import org.eclipse.edc.api.model.QuerySpecDto;
 import org.eclipse.edc.connector.api.management.contractagreement.model.ContractAgreementDto;
 import org.eclipse.edc.web.spi.ApiErrorDetail;
@@ -29,31 +30,19 @@ import org.eclipse.edc.web.spi.ApiErrorDetail;
 import java.util.List;
 
 @OpenAPIDefinition
-@Tag(name = "Contract Agreement - Deprecated")
-@Deprecated(since = "milestone9")
+@Tag(name = "Contract Agreement")
 public interface ContractAgreementApi {
 
     @Operation(description = "Gets all contract agreements according to a particular query",
+            requestBody = @RequestBody(content = @Content(schema = @Schema(implementation = QuerySpecDto.class))),
             responses = {
                     @ApiResponse(responseCode = "200",
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ContractAgreementDto.class)))),
                     @ApiResponse(responseCode = "400", description = "Request body was malformed",
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class))))
-            }, deprecated = true
+            }
     )
-    @Deprecated(since = "milestone9")
-    List<ContractAgreementDto> getAllAgreements(@Valid QuerySpecDto querySpecDto);
-
-    @Operation(description = "Gets all contract agreements according to a particular query",
-            responses = {
-                    @ApiResponse(responseCode = "200",
-                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = ContractAgreementDto.class)))),
-                    @ApiResponse(responseCode = "400", description = "Request body was malformed",
-                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class))))
-            }, deprecated = true
-    )
-    @Deprecated(since = "milestone9")
-    List<ContractAgreementDto> queryAllAgreements(@Valid QuerySpecDto querySpecDto);
+    List<JsonObject> queryAllAgreements(JsonObject querySpecDto);
 
     @Operation(description = "Gets an contract agreement with the given ID",
             responses = {
@@ -63,9 +52,8 @@ public interface ContractAgreementApi {
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class)))),
                     @ApiResponse(responseCode = "404", description = "An contract agreement with the given ID does not exist",
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class))))
-            }, deprecated = true
+            }
     )
-    @Deprecated(since = "milestone9")
-    ContractAgreementDto getContractAgreement(String id);
+    JsonObject getAgreementById(String id);
 
 }
