+++ extensions/control-plane/api/management-api/contract-agreement-api/src/main/java/org/eclipse/edc/connector/api/management/contractagreement/ContractAgreementApiController.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2020 - 2022 Microsoft Corporation
+ *  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,15 +8,13 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - add functionalities
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
  *
  */
 
 package org.eclipse.edc.connector.api.management.contractagreement;
 
-import jakarta.validation.Valid;
-import jakarta.ws.rs.BeanParam;
+import jakarta.json.JsonObject;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
@@ -28,84 +26,78 @@ import org.eclipse.edc.connector.api.management.contractagreement.model.Contract
 import org.eclipse.edc.connector.contract.spi.types.agreement.ContractAgreement;
 import org.eclipse.edc.connector.contract.spi.types.offer.ContractDefinition;
 import org.eclipse.edc.connector.spi.contractagreement.ContractAgreementService;
+import org.eclipse.edc.jsonld.spi.JsonLd;
+import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.query.QuerySpec;
 import org.eclipse.edc.spi.result.Result;
 import org.eclipse.edc.transform.spi.TypeTransformerRegistry;
 import org.eclipse.edc.web.spi.exception.InvalidRequestException;
 import org.eclipse.edc.web.spi.exception.ObjectNotFoundException;
-import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
+import java.util.function.Function;
 
-import static java.lang.String.format;
 import static java.util.Optional.ofNullable;
+import static java.util.stream.Collectors.toList;
 import static org.eclipse.edc.web.spi.exception.ServiceResultHandler.exceptionMapper;
 
 @Produces({ MediaType.APPLICATION_JSON })
-@Path("/contractagreements")
-@Deprecated(since = "milestone9")
+@Path("/v2/contractagreements")
 public class ContractAgreementApiController implements ContractAgreementApi {
-    private final Monitor monitor;
     private final ContractAgreementService service;
+    private final JsonLd jsonLdService;
     private final TypeTransformerRegistry transformerRegistry;
+    private final Monitor monitor;
 
-    public ContractAgreementApiController(Monitor monitor, ContractAgreementService service, TypeTransformerRegistry transformerRegistry) {
-        this.monitor = monitor;
+    public ContractAgreementApiController(ContractAgreementService service, JsonLd jsonLdService, TypeTransformerRegistry transformerRegistry, Monitor monitor) {
         this.service = service;
+        this.jsonLdService = jsonLdService;
         this.transformerRegistry = transformerRegistry;
+        this.monitor = monitor;
     }
 
-    @GET
-    @Override
-    @Deprecated(since = "milestone9")
-    public List<ContractAgreementDto> getAllAgreements(@Valid @BeanParam QuerySpecDto querySpecDto) {
-        return queryContractAgreements(querySpecDto);
-    }
 
     @POST
     @Path("/request")
     @Override
-    @Deprecated(since = "milestone9")
-    public List<ContractAgreementDto> queryAllAgreements(@Valid QuerySpecDto querySpecDto) {
-        return queryContractAgreements(ofNullable(querySpecDto).orElse(QuerySpecDto.Builder.newInstance().build()));
+    public List<JsonObject> queryAllAgreements(JsonObject querySpecDto) {
+
+        Function<Result<JsonObject>, Result<QuerySpec>> expandedMapper =
+                expandedResult -> expandedResult
+                        .compose(jsonObject -> transformerRegistry.transform(jsonObject, QuerySpecDto.class))
+                        .compose(dto -> transformerRegistry.transform(dto, QuerySpec.class));
+
+        var query = ofNullable(querySpecDto)
+                .map(jsonLdService::expand)
+                .map(expandedMapper)
+                .orElse(Result.success(QuerySpec.Builder.newInstance().build()))
+                .orElseThrow(InvalidRequestException::new);
+
+        try (var stream = service.query(query).orElseThrow(exceptionMapper(ContractDefinition.class, null))) {
+            return stream
+                    .map(it -> transformerRegistry.transform(it, ContractAgreementDto.class)
+                            .compose(dto -> transformerRegistry.transform(dto, JsonObject.class))
+                            .compose(jsonLdService::compact))
+                    .peek(r -> r.onFailure(f -> monitor.warning(f.getFailureDetail())))
+                    .filter(Result::succeeded)
+                    .map(Result::getContent)
+                    .collect(toList());
+        }
     }
 
     @GET
     @Path("{id}")
     @Override
-    @Deprecated(since = "milestone9")
-    public ContractAgreementDto getContractAgreement(@PathParam("id") String id) {
-        monitor.debug(format("get contract agreement with ID %s", id));
-
+    public JsonObject getAgreementById(@PathParam("id") String id) {
         return Optional.of(id)
                 .map(service::findById)
-                .map(it -> transformerRegistry.transform(it, ContractAgreementDto.class))
-                .filter(Result::succeeded)
-                .map(Result::getContent)
+                .map(it -> transformerRegistry.transform(it, ContractAgreementDto.class)
+                        .compose(dto -> transformerRegistry.transform(dto, JsonObject.class))
+                        .compose(jsonLdService::compact)
+                        .orElseThrow(failure -> new EdcException(failure.getFailureDetail())))
                 .orElseThrow(() -> new ObjectNotFoundException(ContractAgreement.class, id));
-    }
-
-    @NotNull
-    private List<ContractAgreementDto> queryContractAgreements(QuerySpecDto querySpecDto) {
-        var result = transformerRegistry.transform(querySpecDto, QuerySpec.class);
-        if (result.failed()) {
-            throw new InvalidRequestException(result.getFailureMessages());
-        }
-
-        var spec = result.getContent();
-
-        monitor.debug(format("get all contract agreements from %s", spec));
-
-        try (var stream = service.query(spec).orElseThrow(exceptionMapper(ContractDefinition.class, null))) {
-            return stream
-                    .map(it -> transformerRegistry.transform(it, ContractAgreementDto.class))
-                    .filter(Result::succeeded)
-                    .map(Result::getContent)
-                    .collect(Collectors.toList());
-        }
     }
