+++ extensions/control-plane/api/management-api/contract-definition-api/src/main/java/org/eclipse/edc/connector/api/management/contractdefinition/ContractDefinitionApiExtension.java
@@ -68,7 +68,6 @@ public class ContractDefinitionApiExtension implements ServiceExtension {
 
         var monitor = context.getMonitor();
 
-        webService.registerResource(config.getContextAlias(), new ContractDefinitionApiController(monitor, service, transformerRegistry));
-        webService.registerResource(config.getContextAlias(), new ContractDefinitionNewApiController(jsonLdService, transformerRegistry, service, monitor));
+        webService.registerResource(config.getContextAlias(), new ContractDefinitionApiController(jsonLdService, transformerRegistry, service, monitor));
     }
 }
