+++ extensions/control-plane/api/management-api/transfer-process-api/src/main/java/org/eclipse/edc/connector/api/management/transferprocess/TransferProcessApi.java
@@ -21,8 +21,10 @@ import io.swagger.v3.oas.annotations.links.LinkParameter;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.json.JsonObject;
 import jakarta.validation.Valid;
 import org.eclipse.edc.api.model.IdResponseDto;
 import org.eclipse.edc.api.model.QuerySpecDto;
@@ -36,27 +38,16 @@ import java.util.List;
 
 @OpenAPIDefinition
 @Tag(name = "Transfer Process")
-@Deprecated(since = "milestone9")
 public interface TransferProcessApi {
     @Operation(description = "Returns all transfer process according to a query",
+            requestBody = @RequestBody(content = @Content(schema = @Schema(implementation = QuerySpecDto.class))),
             responses = {
                     @ApiResponse(responseCode = "200",
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = TransferProcessDto.class)))),
                     @ApiResponse(responseCode = "400", description = "Request was malformed",
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class)))) }
     )
-    List<TransferProcessDto> queryAllTransferProcesses(@Valid QuerySpecDto querySpecDto);
-
-    @Operation(description = "Returns all transfer process according to a query",
-            responses = {
-                    @ApiResponse(responseCode = "200",
-                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = TransferProcessDto.class)))),
-                    @ApiResponse(responseCode = "400", description = "Request was malformed",
-                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class)))) },
-            deprecated = true
-    )
-    @Deprecated(since = "milestone8")
-    List<TransferProcessDto> getAllTransferProcesses(@Valid QuerySpecDto querySpecDto);
+    List<JsonObject> queryTransferProcesses(@Valid JsonObject querySpecDto);
 
     @Operation(description = "Gets an transfer process with the given ID",
             responses = {
@@ -68,7 +59,7 @@ public interface TransferProcessApi {
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class))))
             }
     )
-    TransferProcessDto getTransferProcess(String id);
+    JsonObject getTransferProcess(String id);
 
     @Operation(description = "Gets the state of a transfer process with the given ID",
             operationId = "getTransferProcessState",
@@ -81,7 +72,22 @@ public interface TransferProcessApi {
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class))))
             }
     )
-    TransferState getTransferProcessState(String id);
+    JsonObject getTransferProcessState(String id);
+
+    @Operation(description = "Initiates a data transfer with the given parameters. Please note that successfully invoking this endpoint " +
+            "only means that the transfer was initiated. Clients must poll the /{id}/state endpoint to track the state",
+            requestBody = @RequestBody(content = @Content(schema = @Schema(implementation = TransferRequestDto.class))),
+            responses = {
+                    @ApiResponse(responseCode = "200", description = "The transfer was successfully initiated. Returns the transfer process ID and created timestamp",
+                            content = @Content(schema = @Schema(implementation = IdResponseDto.class)),
+                            links = @Link(name = "poll-state", operationId = "getTransferProcessState", parameters = {
+                                    @LinkParameter(name = "id", expression = "$response.body#/id")
+                            })
+                    ),
+                    @ApiResponse(responseCode = "400", description = "Request body was malformed",
+                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class)))),
+            })
+    JsonObject initiateTransferProcess(@Valid JsonObject transferRequestDto);
 
     @Operation(description = "Requests the deprovisioning of resources associated with a transfer process. Due to the asynchronous nature of transfers, a successful " +
             "response only indicates that the request was successfully received. This may take a long time, so clients must poll the /{id}/state endpoint to track the state.",
@@ -97,6 +103,7 @@ public interface TransferProcessApi {
 
     @Operation(description = "Requests the termination of a transfer process. Due to the asynchronous nature of transfers, a successful " +
             "response only indicates that the request was successfully received. Clients must poll the /{id}/state endpoint to track the state.",
+            requestBody = @RequestBody(content = @Content(schema = @Schema(implementation = TerminateTransferDto.class))),
             responses = {
                     @ApiResponse(responseCode = "200", description = "Request to cancel the transfer process was successfully received",
                             links = @Link(name = "poll-state", operationId = "getTransferProcessState")),
@@ -107,32 +114,5 @@ public interface TransferProcessApi {
                     @ApiResponse(responseCode = "409", description = "Could not terminate transfer process, because it is already completed or terminated.",
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class))))
             })
-    void terminateTransferProcess(String id, TerminateTransferDto terminateTransfer);
-
-    @Operation(description = "Requests aborting the transfer process. Due to the asynchronous nature of transfers, a successful " +
-            "response only indicates that the request was successfully received. Clients must poll the /{id}/state endpoint to track the state.",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "Request to cancel the transfer process was successfully received",
-                            links = @Link(name = "poll-state", operationId = "getTransferProcessState")),
-                    @ApiResponse(responseCode = "400", description = "Request was malformed, e.g. id was null",
-                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class)))),
-                    @ApiResponse(responseCode = "404", description = "A contract negotiation with the given ID does not exist",
-                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class))))
-            })
-    @Deprecated(since = "milestone9")
-    void cancelTransferProcess(String id);
-
-    @Operation(description = "Initiates a data transfer with the given parameters. Please note that successfully invoking this endpoint " +
-            "only means that the transfer was initiated. Clients must poll the /{id}/state endpoint to track the state",
-            responses = {
-                    @ApiResponse(responseCode = "200", description = "The transfer was successfully initiated. Returns the transfer process ID and created timestamp",
-                            content = @Content(schema = @Schema(implementation = IdResponseDto.class)),
-                            links = @Link(name = "poll-state", operationId = "getTransferProcessState", parameters = {
-                                    @LinkParameter(name = "id", expression = "$response.body#/id")
-                            })
-                    ),
-                    @ApiResponse(responseCode = "400", description = "Request body was malformed",
-                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiErrorDetail.class)))),
-            })
-    IdResponseDto initiateTransfer(@Valid TransferRequestDto transferRequest);
+    void terminateTransferProcess(String id, JsonObject terminateTransferDto);
 }
