+++ resources/openapi/yaml/management-api/catalog-api.yaml
@@ -1,72 +1,8 @@
 openapi: 3.0.1
 paths:
-  /catalog:
-    get:
-      deprecated: true
-      operationId: getCatalog
-      parameters:
-      - in: query
-        name: providerUrl
-        required: true
-        schema:
-          type: string
-          example: null
-      - in: query
-        name: offset
-        schema:
-          type: integer
-          format: int32
-          example: null
-      - in: query
-        name: limit
-        schema:
-          type: integer
-          format: int32
-          example: null
-      - in: query
-        name: sort
-        schema:
-          type: string
-          enum:
-          - ASC
-          - DESC
-          example: null
-      - in: query
-        name: sortField
-        schema:
-          type: string
-          example: null
-      responses:
-        default:
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Catalog'
-          description: Gets contract offers (=catalog) of a single connector
-      tags:
-      - Catalog (deprecated)
-  /catalog/request:
-    post:
-      deprecated: true
-      operationId: requestCatalog
-      requestBody:
-        content:
-          '*/*':
-            schema:
-              $ref: '#/components/schemas/CatalogRequestDto'
-        required: true
-      responses:
-        default:
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Catalog'
-          description: Gets contract offers (=catalog) of a single connector
-      tags:
-      - Catalog (deprecated)
   /v2/catalog/request:
     post:
-      operationId: requestCatalog_1
+      operationId: requestCatalog
       requestBody:
         content:
           application/json:
