+++ resources/openapi/yaml/management-api/contract-agreement-api.yaml
@@ -1,129 +1,9 @@
 openapi: 3.0.1
 paths:
-  /contractagreements:
-    get:
-      deprecated: true
-      description: Gets all contract agreements according to a particular query
-      operationId: getAllAgreements
-      parameters:
-      - in: query
-        name: offset
-        schema:
-          type: integer
-          format: int32
-          example: null
-      - in: query
-        name: limit
-        schema:
-          type: integer
-          format: int32
-          example: null
-      - in: query
-        name: sort
-        schema:
-          type: string
-          enum:
-          - ASC
-          - DESC
-          example: null
-      - in: query
-        name: sortField
-        schema:
-          type: string
-          example: null
-      responses:
-        "200":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ContractAgreementDto'
-        "400":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: Request body was malformed
-      tags:
-      - Contract Agreement - Deprecated
-  /contractagreements/request:
-    post:
-      deprecated: true
-      description: Gets all contract agreements according to a particular query
-      operationId: queryAllAgreements
-      requestBody:
-        content:
-          '*/*':
-            schema:
-              $ref: '#/components/schemas/QuerySpecDto'
-      responses:
-        "200":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ContractAgreementDto'
-        "400":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: Request body was malformed
-      tags:
-      - Contract Agreement - Deprecated
-  /contractagreements/{id}:
-    get:
-      deprecated: true
-      description: Gets an contract agreement with the given ID
-      operationId: getContractAgreement
-      parameters:
-      - in: path
-        name: id
-        required: true
-        schema:
-          type: string
-          example: null
-      responses:
-        "200":
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ContractAgreementDto'
-          description: The contract agreement
-        "400":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: "Request was malformed, e.g. id was null"
-        "404":
-          content:
-            application/json:
-              schema:
-                type: array
-                example: null
-                items:
-                  $ref: '#/components/schemas/ApiErrorDetail'
-          description: An contract agreement with the given ID does not exist
-      tags:
-      - Contract Agreement - Deprecated
   /v2/contractagreements/request:
     post:
       description: Gets all contract agreements according to a particular query
-      operationId: queryAllAgreements_1
+      operationId: queryAllAgreements
       requestBody:
         content:
           '*/*':
